[Feat] CommunityDetailVC / CommunityDetailView 뷰 분리
